requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.35
Werkzeug==3.0.1
PyJWT==2.8.0
groq>=0.4.1
python-dotenv>=1.0.0
